python-for-ml/P3/customer_segmentation_pipeline/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pyyaml>=6.0
joblib>=1.3.0
plotly>=5.14.0
dash>=2.14.0
dash-bootstrap-components>=1.4.0
scipy>=1.11.0
